Adds Kafka image and TRANSPORT_TYPE to select it with
This adds an image for Kafka/Zookeeper, which can be used to test kafka
integration of zipkin with instrumentation including Finagle, Brave and
HTrace. This exports a new variable to designate the `TRANSPORT_TYPE`,
with choices of `scribe` and `kafka`.

At the moment, there's no Finagle Kafka integraion. I tested this with
an experimental Brave span transport, used in the `zipkin-java` project.

The image bundles together Kafka and Zookeeper using Runit as the entry-
point. This is for simplicitly and to keep the requirements as small as
possible to run tests (ex. we are already several containers). The
approach used is similar to baseimage-docker.

See https://github.com/openzipkin/brave/issues/102
See https://github.com/openzipkin/zipkin/issues/682
See http://phusion.github.io/baseimage-docker