Use overrides for docker-compose mysql setup
The docker-compose-mysql.yml file appears to be incomplete, as it is not
running the collector container. This branch switches that configuration
to be an extention of the default docker-compose.yml file, using
docker-compose overrides, described here:

https://docs.docker.com/compose/extends/#multiple-compose-files

In addition, I've fixed an issue where the zipkin-cassandra container
would not exit if the storage type was not configured correctly. And
I've modified the zipkin-web container to allow overriding the query
service address without requiring linking.