Adds patch to work around Python 2.7.12 incompatibility with cqlsh
It is far from straightforward to install an old version of python in
Alpine as the package repositories only keep current versions.

This has become a problem as newer versions of Python don't work with
out-of-box CQLSH. This adds patches scheduled to go into cassandra,
which allow it to work with newer versions of Python.

See https://issues.apache.org/jira/browse/CASSANDRA-11850