Adds MODULE_OPTS to avoid clobbering 3rd party profile config (#202)
When we switched to JRE 11, we also switched to distroless. There is a
regression where if you set JAVA_OPTS, it undoes any profile setup for
3rd party modules like zipkin-aws. This could be perceived as messages
from SQS not being processed because sqs wasn't loaded.

Instead, this creates a separate env MODULE_OPTS to hold the layered
config. I tested this on zipkin-aws.